Upgrade genesis contract (#154)
