+++ crates/enclave-contract/Cargo.toml
@@ -6,9 +6,6 @@ edition = "2021"
 [dependencies]
 alloy.workspace = true
 anyhow.workspace = true
-serde.workspace = true
-serde_json.workspace = true
-hex.workspace = true
 reqwest.workspace = true
 
 [dev-dependencies]
