+++ crates/enclave-contract/tests/multisig_test.rs
@@ -1,13 +1,22 @@
-use enclave_contract::upgrades_canonical_deploy;
 use enclave_contract::UPGRADE_MULTISIG_ADDRESS;
 use enclave_contract::UPGRADE_OPERATOR_ADDRESS;
 use enclave_contract::{
     can_execute_multisig_proposal, check_proposal_status_v1, create_multisig_proposal,
-    execute_multisig_proposal, get_multisig_vote_count, print_flush, vote_on_multisig_proposal,
+    execute_multisig_proposal, get_multisig_vote_count, vote_on_multisig_proposal,
     ProposalParamsV1, ANVIL_ALICE_SK, ANVIL_BOB_SK,
 };
 use std::thread::sleep;
 use std::time::Duration;
+
+/// Prints a string to standard output and immediately flushes the output buffer.
+/// Useful to see prints immediately during long-running Cargo tests.
+pub fn print_flush<S: AsRef<str>>(s: S) {
+    use std::io::Write;
+    let stdout = std::io::stdout();
+    let mut handle = stdout.lock(); // lock ensures safe writing
+    write!(handle, "{}", s.as_ref()).unwrap();
+    handle.flush().unwrap();
+}
 
 /// Test the complete multisig workflow for controlling UpgradeOperator
 /// This test verifies that the multisig contract can properly control the UpgradeOperator
@@ -15,17 +24,12 @@ use std::time::Duration;
 #[tokio::test(flavor = "multi_thread")]
 pub async fn test_multisig_upgrade_operator_workflow() -> Result<(), anyhow::Error> {
     // Set path to the factory contract's json file
-    let factory_json_path = "contracts/out/UpgradeOperatorFactory.sol/UpgradeOperatorFactory.json";
     let reth_rpc = "http://localhost:8545";
     let multisig_address = UPGRADE_MULTISIG_ADDRESS
         .parse::<alloy::primitives::Address>()
         .unwrap();
     let upgrade_operator_address = UPGRADE_OPERATOR_ADDRESS
         .parse::<alloy::primitives::Address>()
-        .unwrap();
-
-    upgrades_canonical_deploy(factory_json_path, reth_rpc)
-        .await
         .unwrap();
 
     // Wait a bit for the transaction to be processed
