+++ crates/enclave-server/tests/integration/main.rs
@@ -6,3 +6,5 @@ mod snapshot;
 
 #[cfg(test)]
 mod booter;
+
+mod utils;
